Refactor post route exports for Next.js compliance
